Make number answer type work
Copied all the old code in, and made it work with the new system. I did
have to separate out the transformers from the examples, which could be
confusing if we try to add more types, but it should work well for now.

Also, inheritance between number answer types is a bit broken for now.