Remove reference to non-existant "B" variable
In the a-o-xmk and a-o-xmk2 answer-types of limits_2, there's a "B"
variable that isn't defined, but it renders fine without it (and the
question/answer still makes sense). So, if not running in debug mode,
the exception is caught and ignored, and the error goes unseen. This
removes all references to "B", which fixes everything.